+++ src/index.ts
@@ -16,6 +16,14 @@ import triggerMembershipCancelled from "./triggers/triggerMembershipCancelled";
 import triggerMembershipCancellationDateReached from "./triggers/triggerMembershipCancellationDateReached";
 import triggerDropInPassPurchased from "./triggers/triggerDropInPassPurchased";
 import triggerExternalBookingApproved from "./triggers/triggerExternalBookingApproved";
+import triggerBookingUpdated from "./triggers/triggerBookingUpdated";
+import triggerExternalBookingUpdated from "./triggers/triggerExternalBookingUpdated";
+import triggerBookingDeleted from "./triggers/triggerBookingDeleted";
+import triggerMembershipUncanceled from "./triggers/triggerMembershipUncanceled";
+import triggerMembershipReactivated from "./triggers/triggerMembershipReactivated";
+import triggerMembershipDeleted from "./triggers/triggerMembershipDeleted";
+import triggerMembershipUpdated from "./triggers/triggerMembershipUpdated";
+import triggerMembershipPlanChangedEvent from "./triggers/triggerMembershipPlanChangedEvent";
 
 const { version } = require("../package.json");
 
@@ -43,9 +51,17 @@ export default {
     [triggerInvoiceCreated.key]: triggerInvoiceCreated,
     [triggerExternalBookingApproved.key]: triggerExternalBookingApproved,
     [triggerExternalBookingCreated.key]: triggerExternalBookingCreated,
+    [triggerExternalBookingUpdated.key]: triggerExternalBookingUpdated,
     [triggerDropInPassPurchased.key]: triggerDropInPassPurchased,
     [triggerMembershipCancellationDateReached.key]:
       triggerMembershipCancellationDateReached,
+    [triggerBookingUpdated.key]: triggerBookingUpdated,
+    [triggerBookingDeleted.key]: triggerBookingDeleted,
+    [triggerMembershipUncanceled.key]: triggerMembershipUncanceled,
+    [triggerMembershipReactivated.key]: triggerMembershipReactivated,
+    [triggerMembershipDeleted.key]: triggerMembershipDeleted,
+    [triggerMembershipUpdated.key]: triggerMembershipUpdated,
+    [triggerMembershipPlanChangedEvent.key]: triggerMembershipPlanChangedEvent,
     // Lists for dropdowns
     [getSubdomains.key]: getSubdomains,
   },