+++ src/triggers/triggerBookingDeleted.ts
@@ -1,16 +1,21 @@
 import { ZObject } from "zapier-platform-core";
 import { KontentBundle } from "../types/kontentBundle";
-import { subscribeHook, unsubscribeHook } from "../utils/api";
+import {
+  subscribeHook,
+  unsubscribeHook,
+  getResource,
+  getMembership,
+} from "../utils/api";
 import { getSubdomainField } from "../fields/getSudomainsField";
 import { SubscribeBundleInputType } from "../types/subscribeType";
 import { bookingSample } from "../utils/samples";
 import { BookingOutput } from "../types/outputs";
+import { apiResponseToBookingOutput } from "../utils/api-to-output";
 import { HookTrigger } from "../types/trigger";
-import { loadBookingAndConvertToOutput } from "../utils/load-to-output";
 import { listRecentBookingsAndConvertToOutput } from "../utils/list";
 
-const hookLabel = "Booking Created";
-const event = "created_booking";
+const hookLabel = "Booking Deleted";
+const event = "deleted_booking";
 
 async function subscribeHookExecute(
   z: ZObject,
@@ -34,12 +39,22 @@ async function parsePayload(
   z: ZObject,
   bundle: KontentBundle<{}>,
 ): Promise<BookingOutput[]> {
-  if (bundle.cleanedRequest) {
-    const bookingId = bundle.cleanedRequest.url.split("/").pop();
-    return loadBookingAndConvertToOutput(z, bundle, bookingId);
-  } else {
-    return [];
+  if (bundle.cleanedRequest?.booking) {
+    const booking = bundle.cleanedRequest.booking;
+    const subdomain = (bundle.inputData as any).subdomain as string;
+    const membershipId = booking.relationships?.membership?.data?.id;
+    const resourceId = booking.relationships.resource.data?.id;
+    const resource = await getResource(z, resourceId);
+    if (!resource) {
+      return [];
+    }
+    let membership = null;
+    if (membershipId && subdomain) {
+      membership = await getMembership(z, subdomain, membershipId);
+    }
+    return [apiResponseToBookingOutput(booking, membership, resource)];
   }
+  return [];
 }
 
 const trigger: HookTrigger = {
@@ -47,7 +62,7 @@ const trigger: HookTrigger = {
   noun: hookLabel,
   display: {
     label: hookLabel,
-    description: "Triggers when a booking is made.",
+    description: "Triggers when a booking is deleted.",
   },
   operation: {
     type: "hook",