+++ src/triggers/triggerExternalBookingUpdated.ts
@@ -1,16 +1,20 @@
 import { ZObject } from "zapier-platform-core";
 import { KontentBundle } from "../types/kontentBundle";
-import { subscribeHook, unsubscribeHook } from "../utils/api";
-import { getSubdomainField } from "../fields/getSudomainsField";
+import {
+  getExternalBookingFromBookingId,
+  listRecentExternalBookings,
+  subscribeHook,
+  unsubscribeHook,
+} from "../utils/api";
 import { SubscribeBundleInputType } from "../types/subscribeType";
-import { bookingSample } from "../utils/samples";
-import { BookingOutput } from "../types/outputs";
+import { getSubdomainField } from "../fields/getSudomainsField";
+import { externalBookingSample } from "../utils/samples";
+import { ExternalBookingOutput } from "../types/outputs";
+import { apiResponseToExternalBookingOutput } from "../utils/api-to-output";
 import { HookTrigger } from "../types/trigger";
-import { loadBookingAndConvertToOutput } from "../utils/load-to-output";
-import { listRecentBookingsAndConvertToOutput } from "../utils/list";
 
-const hookLabel = "Booking Created";
-const event = "created_booking";
+const hookLabel = "External Booking Updated";
+const event = "updated_booking";
 
 async function subscribeHookExecute(
   z: ZObject,
@@ -33,21 +37,22 @@ async function unsubscribeHookExecute(
 async function parsePayload(
   z: ZObject,
   bundle: KontentBundle<{}>,
-): Promise<BookingOutput[]> {
-  if (bundle.cleanedRequest) {
-    const bookingId = bundle.cleanedRequest.url.split("/").pop();
-    return loadBookingAndConvertToOutput(z, bundle, bookingId);
+): Promise<ExternalBookingOutput[]> {
+  const bookingId = bundle.cleanedRequest.url.split("/").pop();
+  const response = await getExternalBookingFromBookingId(z, bookingId);
+  if (response) {
+    return [apiResponseToExternalBookingOutput(response)];
   } else {
     return [];
   }
 }
 
 const trigger: HookTrigger = {
-  key: event,
+  key: `${event}_external`,
   noun: hookLabel,
   display: {
     label: hookLabel,
-    description: "Triggers when a booking is made.",
+    description: "Triggers when an external booking is updated.",
   },
   operation: {
     type: "hook",
@@ -61,10 +66,12 @@ const trigger: HookTrigger = {
     performList: async (
       z: ZObject,
       bundle: KontentBundle<SubscribeBundleInputType>,
-    ): Promise<BookingOutput[]> => {
-      return listRecentBookingsAndConvertToOutput(z, bundle);
+    ): Promise<ExternalBookingOutput[]> => {
+      const bookings = await listRecentExternalBookings(z, bundle);
+      return bookings.map(apiResponseToExternalBookingOutput);
     },
-    sample: bookingSample,
+
+    sample: externalBookingSample,
   },
 };
 export default trigger;